Address review: move to browser
