+++ Lib/idlelib/idle_test/test_browser.py
@@ -5,6 +5,7 @@ from test.support import requires
 import unittest
 from unittest import mock
 from idlelib.idle_test.mock_idle import Func
+from idlelib.util import py_extensions
 
 from collections import deque
 import os.path
@@ -56,6 +57,14 @@ class ModuleBrowserTest(unittest.TestCase):
         self.assertTrue(mb.top.destroy.called)
         self.assertTrue(mb.node.destroy.called)
         del mb.top.destroy, mb.node.destroy
+
+    def test_is_browseable_extension(self):
+        for ext in py_extensions:
+            with self.subTest(ext=ext):
+                filename = f'{self.path}{ext}'
+                actual = browser.is_browseable_extension(filename)
+                expected = ext not in browser.browseable_extension_blocklist
+                self.assertEqual(actual, expected)
 
 
 # Nested tree same as in test_pyclbr.py except for supers on C0. C1.
