+++ Lib/idlelib/idle_test/test_util.py
@@ -19,13 +19,11 @@ class UtilTest(unittest.TestCase):
                 filename = f"{self.path}{ext}"
                 self.assertFalse(util.is_supported_extension(filename))
 
-    def test_is_browseable_extension(self):
-        for ext in util.py_extensions:
-            with self.subTest(ext=ext):
-                filename = f'{self.path}{ext}'
-                actual = util.is_browseable_extension(filename)
-                expected = ext not in util.browseable_extension_blocklist
-                self.assertEqual(actual, expected)
+    def test_get_ext(self):
+        for path in "/a/b/c.py", "d.py", "a.py.py":
+            with self.subTest(path=path):
+                ext = util.get_ext(path)
+                self.assertEqual(ext, ".py")
 
 
 if __name__ == '__main__':
